refactor: move ssh
